Updates for rtp and udp directories of gst-plugins
Original commit message from CVS:
Updates for rtp and udp directories of gst-plugins

UDP: adds a time to live Element Property to udpsink, fixes comments in udpsrc
RTP:
1) fairly complete rewrite of the encoding and parsing code, encoder now
passes analysis by ethereal and can play to the parser and to mplayer.
2) Adds support for configurable dynamic payload processing using a new element
property "rtpmap" which is a string of rtpmap records in the same format as
would be present in an SDP record.
3) Adds support for big and little endian peers on the pads
4) Adds support for mtu element property on rtpL16enc and will split incoming
buffers to avoid IP layer fragmentation.