ext/faad/gstfaad.c: I don't understand how this got unnoticed for so long - did the faad API change or something ? al...
Original commit message from CVS:

* ext/faad/gstfaad.c: (gst_faad_sinkconnect), (gst_faad_chain):
I don't understand how this got unnoticed for so long - did the
faad API change or something ?
also, we're getting samplerate and channels in the first call,
but not actually *doing* anything with that.  Sigh.
* ext/libmng/gstmngdec.c:
switch from 0.0 to 1.0 minimum framerate, even though that's
*stupid* because we want to use mngdec for things that happen
less than once a second, but it's the only way to have
somewhat acceptable negotiation.
* ext/theora/theoraenc.c: (theora_enc_chain):
add a debug message.