uvch264src: don't error out on incomplete aux data segment
It appears that the Logitech C920 sometimes drops the next
to last segment of RAW aux data contained within the MJPEG
container.  H264 data that is multiplexed with in the same
container does not appear to be affected.  This appears to
be a bug in the Logitech C920 firmware and uvch264src should
not error out in this case.

Sometimes it can take 24 hours of continuous streaming for
the problem to occur, but sometimes it takes only a couple
of hours.

https://bugzilla.gnome.org/show_bug.cgi?id=706276